ISSUE #4780 - fix after merge issues
