+++ frontend/src/v5/services/api/drawings.ts
@@ -32,7 +32,7 @@ export const removeFavourite = (teamspace, projectId, drawingId): Promise<AxiosR
 
 const categories =  ['A drawing category', 'Another drawing category', 'Yet another one'];
 
-const arr = (new Array(1000)).fill(0);
+const arr = (new Array(10)).fill(0);
 
 const drawings = arr.map((_, index) => {
 	return {
@@ -68,20 +68,19 @@ const stats = arr.map((_, index) => {
 
 	const calibration = total ? randCal() : CalibrationStates.EMPTY;
 	const latestRevision = total ? 'Revision ' + total : undefined;
-	const status = total ? 'status' : undefined;
+	const status = total ? DrawingUploadStatus.OK : undefined;
 
 	return {
 		_id: drawings[index]._id,
 		revisions : {
 			lastUpdated,
-			drawingNumber: uuid(), 
 			total,
-			calibration,
-			category: categories[Math.round(Math.random() * (categories.length - 1))],
 			latestRevision,
-			status,
 		},
-		status: DrawingUploadStatus.OK,
+		drawingNumber: uuid(), 
+		calibration,
+		category: categories[Math.round(Math.random() * (categories.length - 1))],
+		status,
 	} as DrawingStats;
 });
